requirements.txt
flask==3.0.0
pandas==2.1.4
Pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
google-generativeai==0.3.2